Create kernel and acq module
